Add bundled demo kernels (holoso.demos)
Function-form demo kernels (dot2, madd, cube, poly3, blend, ekf_update)
ship as package data with a load_demos() reader, so any holoso install --
including the in-browser Pyodide wheel used by holoso-web -- can list and
load them via importlib.resources, with no separate corpus.

ekf_update is now the single source for the EKF kernel; examples/ekf1.py
imports it instead of restating the body.
