+++ holoso/demos/blend.py
@@ -1,0 +1,5 @@
+"""Floating-point division followed by a scaled add -- exercises the divide operator."""
+
+
+def blend(a, b, c):
+    return a / b + c * 2.0
