+++ holoso/demos/cube.py
@@ -1,0 +1,5 @@
+"""Integer power: a**3 lowers to a chain of multiplications."""
+
+
+def cube(a):
+    return a**3
